+++ src/Mac.Digital/Services/IPowerService.cs
@@ -24,7 +24,7 @@ namespace Mac.Digital.Services
         /// <summary>
         /// Gets an observable value indicating the power usage in Watts.
         /// </summary>
-        IObservable<double> PowerInWatts { get; }
+        IObservable<decimal> PowerInWatts { get; }
 
         /// <summary>
         /// Turns the machine on.
